requirements.txt
bert-for-tf2==0.14.1
gdown==3.6.4
pandas==1.0.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
sentencepiece==0.1.85
sklearn==0.0
tensorflow==2.2.0rc2
tqdm==4.38.0
